refactor: mapper 함수 별도 파일로 분리
